Вынес некоторые функции в util.js. Почистил модули от лишнего
